pyproject.toml
[tool.poetry]
name = "task2(o2rus)"
version = "0.1.0"
description = ""
authors = ["korpaxdev <korpaxdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.14.0"
black = "^24.2.0"
pydantic-settings = "^2.2.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.3.6"
redis = "^5.0.3"
pydantic = {extras = ["email"], version = "^2.6.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
